Simplify the files for OpenType
Editing the font specific .fea files helped, but did not take address all
the differences in rendering. In one test, the OpenType code in the
master.fea may well be better than in master.sfd at positioning.

The file master.fea is being returned to how it was originally,
the differences are due to whitespace normalization.