Use a more UFO centric build
The SFD files that were removed were going to be the new source
files at one point. They are currently generated by the ufo script,
and then used to generate the UFO sources that are used to build
the fonts. Eventually, the UFO sources will be checked in,
and the ufo script will not need to be run to generate sources.