Forgot to include a sub form of ra
Added kerning (that was in the last commit) with ra.below.ra. The glyphs ra.below and ra.below.large are posistioned directly under the base consonant and therefore do not need to be kerned. The glyph ra.below.subcons is only for a second level sub form (the other glyphs mentioned here are first level sub forms).